pyproject.toml
[project]
name = "FBDS_Downloader"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "anderson", email = "anderson@giscaruso.com.br"},
]
dependencies = [
    "pytest>=8.2.0",
    "requests>=2.31.0",
    "patool>=2.2.0",
    "tomli>=2.0.1",
    "geopandas>=0.14.4",
    "pyshp>=2.3.1",
    "ipykernel>=6.29.4",
    "fastapi==0.108.0",
    "uvicorn>=0.29.0",
    "nicegui>=1.4.9",
    "bs4>=0.0.2",
    "redmail>=0.6.0",
    "setuptools>=70.2.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
